pyproject.toml
[project]
name = "fastapi_course"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "alembic==1.14.1",
    "annotated-types==0.7.0",
    "anyio==4.8.0",
    "asttokens==3.0.0",
    "bcrypt==4.2.1",
    "certifi==2025.1.31",
    "cffi==1.17.1",
    "click==8.1.8",
    "colorama==0.4.6",
    "comm==0.2.2",
    "cryptography==44.0.1",
    "debugpy==1.8.12",
    "decorator==5.1.1",
    "dnspython==2.7.0",
    "ecdsa==0.19.0",
    "email-validator==2.2.0",
    "executing==2.2.0",
    "fastapi==0.115.8",
    "fastapi-cli==0.0.7",
    "greenlet==3.1.1",
    "h11==0.14.0",
    "httpcore==1.0.7",
    "httptools==0.6.4",
    "httpx==0.28.1",
    "idna==3.10",
    "ipykernel==6.29.5",
    "ipython==8.32.0",
    "itsdangerous==2.2.0",
    "jedi==0.19.2",
    "jinja2==3.1.5",
    "jupyter-client==8.6.3",
    "jupyter-core==5.7.2",
    "mako==1.3.9",
    "markdown-it-py==3.0.0",
    "markupsafe==3.0.2",
    "matplotlib-inline==0.1.7",
    "mdurl==0.1.2",
    "nest-asyncio==1.6.0",
    "numpy==2.2.2",
    "orjson==3.10.15",
    "packaging==24.2",
    "parso==0.8.4",
    "passlib==1.7.4",
    "platformdirs==4.3.6",
    "prompt-toolkit==3.0.50",
    "psutil==7.0.0",
    "psycopg-binary==3.2.4",
    "psycopg2-binary>=2.9.10",
    "pure-eval==0.2.3",
    "pyasn1==0.6.1",
    "pycparser==2.22",
    "pydantic==2.10.6",
    "pydantic-core==2.27.2",
    "pydantic-extra-types==2.10.2",
    "pydantic-settings==2.7.1",
    "pygments==2.19.1",
    "python-dateutil==2.9.0.post0",
    "python-dotenv==1.0.1",
    "python-jose==3.3.0",
    "python-multipart==0.0.20",
    "pyyaml==6.0.2",
    "pyzmq==26.2.1",
    "rich==13.9.4",
    "rich-toolkit==0.13.2",
    "rsa==4.9",
    "shellingham==1.5.4",
    "six==1.17.0",
    "sniffio==1.3.1",
    "sqlalchemy==1.4.28",
    "stack-data==0.6.3",
    "starlette==0.45.3",
    "tornado==6.4.2",
    "traitlets==5.14.3",
    "typer==0.15.1",
    "typing-extensions==4.12.2",
    "tzdata==2025.1",
    "ujson==5.10.0",
    "uvicorn==0.34.0",
    "watchfiles==1.0.4",
    "wcwidth==0.2.13",
    "websockets==14.2",
]
